deleteById

+++ src/main/java/com/senior/dreamteam/services/UserService.java
@@ -116,9 +116,10 @@ public class UserService {
         try {
             Optional<User> userOptional = userRepository.findById(id);
             if (userOptional.isPresent()) {
-                if((jwtTokenUtil.isTokenValid(token, userOptional.get()) && jwtTokenUtil.isAccessToken(token)) || jwtTokenUtil.isAdminToken(token))
-                userRepository.deleteById(id);
-                return "User removed successfully";
+                if ((jwtTokenUtil.isTokenValid(token, userOptional.get()) && jwtTokenUtil.isAccessToken(token)) && jwtTokenUtil.isAdminToken(token)) {
+                    userRepository.deleteById(id);
+                    return "User removed successfully";
+                }
             } else {
                 throw new DemoGraphqlException("User not found with id: " + id);
             }
